Emulate the non-linear torque of the robot

+++ Assets/Scripts/Utils.cs
@@ -1,7 +1,7 @@
 using System;
 
 public static class Utils {
-    public static float Map(int value, int fromLow, int fromHigh, int toLow, int toHigh)
+    public static float Map(int value, int fromLow, int fromHigh, float toLow, float toHigh)
     {
         return (float)(value - fromLow) * (toHigh - toLow) / (fromHigh - fromLow) + toLow;
     }
@@ -10,7 +10,7 @@ public static class Utils {
         return Math.Max(Math.Min(value, max), min);
     }
 
-    public static float MapAndLimit(int value, int fromLow, int fromHigh, int toLow, int toHigh)
+    public static float MapAndLimit(int value, int fromLow, int fromHigh, float toLow, float toHigh)
     {
         return Limit(Map(value, fromLow, fromHigh, toLow, toHigh), toLow, toHigh);
     }
